move main.rs into the db submodule
