+++ docbrown/core/Cargo.toml
@@ -9,13 +9,12 @@ keywords = ["graph", "temporal-graph", "temporal"]
 [dependencies]
 roaring="0.10"
 itertools="0.10"
-csv = "1.1.6"
+# csv = "1.1.6"
 replace_with = "0.1"
 thiserror = "1"
 rand = "0.8.5"
-flate2 = "1.0"
 sorted_vector_map = "0.1"
-flume = "0.10"
+# flume = "0.10"
 futures = "0.3"
 genawaiter = "0.99"
 parking_lot = "0.12"
@@ -28,6 +27,7 @@ parking_lot = "0.12"
 criterion = "0.4"
 quickcheck = "1"
 quickcheck_macros = "1"
+csv = "1.1.6"
 
 [[bench]]
 name = "tgraph_benchmarks"
