+++ docbrown/core/src/lib.rs
@@ -2,7 +2,6 @@
 #[macro_use(quickcheck)]
 extern crate quickcheck_macros;
 mod bitset;
-pub mod db;
 pub mod graph;
 pub mod lsm;
 mod adj;
