+++ docbrown/db/Cargo.toml
@@ -13,6 +13,8 @@ flume = "0.10"
 futures = {version = "0.3", features = ["thread-pool"] }
 replace_with = "0.1"
 itertools="0.10"
+csv = "1.1.6"
+flate2 = "1.0"
 
 [dev-dependencies]
 quickcheck = "1"
