+++ docbrown/db/src/lib.rs
@@ -2,11 +2,12 @@
 #[macro_use(quickcheck)]
 extern crate quickcheck_macros;
 
-use std::{sync::Arc, rc::Rc};
+use std::sync::Arc;
 
 use docbrown_core::{graph::TemporalGraph, Prop};
 use parking_lot::RwLock;
 
+#[derive(Debug, Clone)]
 pub struct GraphDB {
     nr_shards: usize,
     shards: Vec<Arc<RwLock<TemporalGraph>>>,
@@ -33,7 +34,7 @@ impl GraphDB {
         }
     }
 
-    pub fn add_vertex(&self, v: u64, t: u64, props: Vec<Prop>) -> () {
+    pub fn add_vertex(&self, v: u64, t: u64, props: Vec<Prop>) {
         let shard_id = self.shard_from_global_vid(v);
 
         self.write_shard(shard_id, move |shard| shard.add_vertex_props(v, t, &props));
@@ -96,6 +97,13 @@ impl GraphDB {
 #[cfg(test)]
 mod db_tests {
     use itertools::Itertools;
+    use csv::StringRecord;
+
+    use std::{
+        collections::hash_map::DefaultHasher,
+        hash::{Hash, Hasher},
+        path::PathBuf,
+    };
 
     use super::*;
 
@@ -111,4 +119,45 @@ mod db_tests {
         assert_eq!(g.len(), expected_len)
 
     }
+
+    #[test]
+    fn db_lotr() {
+
+        let g = GraphDB::new(4);
+
+        fn calculate_hash<T: Hash>(t: &T) -> u64 {
+            let mut s = DefaultHasher::new();
+            t.hash(&mut s);
+            s.finish()
+        }
+
+        fn parse_record(rec: &StringRecord) -> Option<(String, String, u64)> {
+            let src = rec.get(0).and_then(|s| s.parse::<String>().ok())?;
+            let dst = rec.get(1).and_then(|s| s.parse::<String>().ok())?;
+            let t = rec.get(2).and_then(|s| s.parse::<u64>().ok())?;
+            Some((src, dst, t))
+        }
+
+        let lotr_csv: PathBuf = [env!("CARGO_MANIFEST_DIR"), "resources/test/lotr.csv"]
+            .iter()
+            .collect();
+
+        let empty:Vec<(String, Prop)> = vec![];
+
+        if let Ok(mut reader) = csv::Reader::from_path(lotr_csv) {
+            for rec_res in reader.records() {
+                if let Ok(rec) = rec_res {
+                    if let Some((src, dst, t)) = parse_record(&rec) {
+                        let src_id = calculate_hash(&src);
+                        let dst_id = calculate_hash(&dst);
+
+                        g.add_vertex(src_id, t, vec![]);
+                        g.add_vertex(dst_id, t, vec![]);
+                        g.add_edge(src_id, dst_id, t, &empty);
+                    }
+                }
+            }
+        }
+    }
+
 }
